requirements.txt
fastapi
uvicorn
websockets
python-dotenv
google-genai
audioop-lts
